// api/src/utils/status-code.ts
export enum STATUS_CODE {
    OK = 200,
    CREATED = 201,
    UPDATED = 200,
    NO_CONTENT = 204,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    INTERNAL_SERVER_ERROR = 500,
    SERVICE_UNAVAILABLE = 503,
    DEFAULT_ERROR = 418,
}

export default STATUS_CODE;